+++ mira/dkg/model.py
@@ -38,7 +38,7 @@ from mira.sources.askenet.flux_span import reproduce_ode_semantics, \
 from mira.sources.askenet.petrinet import template_model_from_askenet_json
 from mira.sources.bilayer import template_model_from_bilayer
 from mira.sources.biomodels import get_sbml_model
-from mira.sources.ascet.petri import template_model_from_petri_json
+from mira.sources.acset.petri import template_model_from_petri_json
 from mira.sources.sbml import template_model_from_sbml_string
 
 __all__ = [
