+++ tests/test_model_api.py
@@ -28,7 +28,7 @@ from mira.modeling.viz import GraphicalModel
 from mira.sources.askenet.petrinet import template_model_from_askenet_json
 from mira.sources.bilayer import template_model_from_bilayer
 from mira.sources.biomodels import get_sbml_model
-from mira.sources.ascet.petri import template_model_from_petri_json
+from mira.sources.acset.petri import template_model_from_petri_json
 from mira.sources.sbml import template_model_from_sbml_string
 from tests import sorted_json_str, remove_all_sympy
 import requests
